requirement.txt
Flask==2.3.3
Pillow==10.0.0